pyproject.toml
[tool.poetry]
name = "niwanowa-rssreader"
version = "0.1.0"
description = ""
authors = ["niwanowa <40478818+niwanowa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
discord-py = "^2.3.2"
python-dotenv = "^1.0.0"
feedparser = "^6.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
